Server side exceptions handled

+++ src/main/java/com/scm/ftn/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.scm.ftn.student;
 
+import com.scm.ftn.student.exceptions.BadRequestException;
+import com.scm.ftn.student.exceptions.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -18,12 +20,19 @@ public class StudentService {
 
     public void addStudent(Student student) {
         //check if email is taken
-
+        Boolean existsEmail = studentRepository.selectExistsEmail(student.getEmail());
+        if(existsEmail){
+            throw new BadRequestException("Email " + student.getEmail() + " already exist");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
         //check if student exist
+
+        if(!studentRepository.existsById(studentId)){
+            throw new StudentNotFoundException("Student with id " + studentId + "does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
